Moved visualizations.py to the Kalpana folder, final updates on export examples
